Workaround for Android Gradle plugin issue when multiple NDKs are available. (#13)
The Android Gradle plugin version < 4.0 fails to find the NDK if there is more than 1 available.
Recently Github added an additional NDK to the action runner's base image, so our builds are
now failing. This change adds a step to the action to move the additional NDKs out of Gradle's
sight, which fixes our builds.

The script is not executable and has additional checks to ensure it's not accidentally run or
sourced outside of a Github action.